move non_matching_operators to operators, remove obsolete code, move things to utils_conga_2d
